# Slightly speed up IDependencyCollection