Changes some `IReadonlyDictionary` arguments to `FrozenDictionay`, which seems to make things slightly faster.

Benchmarks for resolving 3 dependencies:

Old master:
|              Method |     Mean |   Error |  StdDev |
|-------------------- |---------:|--------:|--------:|
| ResolveDependencies | 132.0 ns | 2.65 ns | 7.70 ns |


a200d73ef94e671deed5489b7c51382847ee6cb4:
|              Method |     Mean |    Error |   StdDev |
|-------------------- |---------:|---------:|---------:|
| ResolveDependencies | 96.82 ns | 1.732 ns | 3.538 ns |

This PR:
|              Method |     Mean |    Error |   StdDev |
|-------------------- |---------:|---------:|---------:|
| ResolveDependencies | 89.01 ns | 1.805 ns | 4.184 ns |